requirements.txt
accelerate==0.31.0
certifi==2023.11.17
charset-normalizer==3.3.2
clean-fid==0.1.35
colorama==0.4.4
filelock==3.13.1
fsspec==2023.10.0
huggingface-hub==0.21.3
idna==3.6
importlib-metadata==6.11.0
Jinja2==3.1.3
MarkupSafe==2.1.3
mpmath==1.3.0
networkx==3.2.1
ninja==1.11.1.1
notebook==6.5.4
notebook_shim==0.2.3
numba==0.58.1
numpy==1.26.4
opencv-python==4.10.0.82
packaging==21.3
peft==0.11.1
pillow==10.2.0
pip==23.3.2
psutil==5.9.7
PyYAML==6.0.1
pyzmq==25.1.2
regex==2023.12.25
requests==2.31.0
safetensors==0.4.3
scipy==1.11.4
setuptools==69.0.3
sympy==1.12
tokenizers==0.15.2
torch==2.3.1
torchmetrics==1.3.1
torchvision==0.18.1
torch-fidelity==0.3.0
tqdm==4.66.1
transformers==4.38.2
triton==2.3.1
typing_extensions==4.9.0
urllib3==1.26.18
wcwidth==0.2.13
wheel==0.42.0
zipp==3.17.0
